+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
 require 'minitest/autorun'
 require 'byebug'
 require 'middleman'
-require 'ffaker'
+require 'faker'